Extract StreamEventHandler from StreamObserverService (#1781)
Separate event handling logic into StreamEventHandler struct,
reducing StreamObserverService init from 11 to 3 dependencies.
