+++ Gem/Services/ServicesFactory.swift
@@ -176,7 +176,7 @@ struct ServicesFactory {
             priceUpdater: streamSubscriptionService,
             preferences: preferences
         )
-        let streamObserverService = Self.makeStreamObserverService(
+        let streamEventService = StreamEventService(
             walletStore: storeManager.walletStore,
             notificationStore: storeManager.inAppNotificationStore,
             priceService: priceService,
@@ -185,8 +185,11 @@ struct ServicesFactory {
             transactionsService: transactionsService,
             nftService: nftService,
             perpetualService: perpetualService,
+            preferences: preferences
+        )
+        let streamObserverService = StreamObserverService(
             subscriptionService: streamSubscriptionService,
-            preferences: preferences,
+            eventService: streamEventService,
             webSocket: webSocket
         )
         let explorerService = ExplorerService.standard
@@ -643,31 +646,4 @@ extension ServicesFactory {
         return WebSocketConnection(configuration: configuration)
     }
 
-    private static func makeStreamObserverService(
-        walletStore: WalletStore,
-        notificationStore: InAppNotificationStore,
-        priceService: PriceService,
-        priceAlertService: PriceAlertService,
-        balanceUpdater: any BalanceUpdater,
-        transactionsService: TransactionsService,
-        nftService: NFTService,
-        perpetualService: any HyperliquidPerpetualServiceable,
-        subscriptionService: StreamSubscriptionService,
-        preferences: Preferences,
-        webSocket: any WebSocketConnectable
-    ) -> StreamObserverService {
-        StreamObserverService(
-            walletStore: walletStore,
-            notificationStore: notificationStore,
-            priceService: priceService,
-            priceAlertService: priceAlertService,
-            balanceUpdater: balanceUpdater,
-            transactionsService: transactionsService,
-            nftService: nftService,
-            perpetualService: perpetualService,
-            subscriptionService: subscriptionService,
-            preferences: preferences,
-            webSocket: webSocket
-        )
-    }
 }
